[CHORE] 패키지 구조를 변경하였습니다.
