+++ app/src/main/AndroidManifest.xml
@@ -16,7 +16,7 @@
         android:usesCleartextTraffic="true"
         tools:targetApi="31">
         <activity
-            android:name=".SigninActivity"
+            android:name=".ui.auth.SigninActivity"
             android:screenOrientation="portrait"
             android:exported="true">
             <intent-filter>
@@ -25,11 +25,11 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".SignupActivity"
+            android:name=".ui.auth.SignupActivity"
             android:screenOrientation="portrait"
             android:exported="false" />
         <activity
-            android:name=".HomeActivity"
+            android:name=".ui.main.MainActivity"
             android:configChanges="keyboardHidden|orientation|screenSize"
             android:exported="false" />
     </application>
