+++ app/src/main/java/org/sopt/dosopttemplate/FriendViewHolder.kt
@@ -4,6 +4,7 @@ import androidx.recyclerview.widget.RecyclerView
 import coil.load
 import coil.transform.RoundedCornersTransformation
 import org.sopt.dosopttemplate.databinding.ItemFriendBinding
+import org.sopt.dosopttemplate.domain.model.Friend
 
 
 class FriendViewHolder(
