+++ app/src/main/java/org/sopt/dosopttemplate/ui/auth/SignUpActivity.kt
@@ -1,4 +1,4 @@
-package org.sopt.dosopttemplate
+package org.sopt.dosopttemplate.ui.auth
 
 import android.content.Intent
 import android.os.Bundle
@@ -6,8 +6,16 @@ import androidx.activity.OnBackPressedCallback
 import androidx.appcompat.app.AlertDialog
 import androidx.appcompat.app.AppCompatActivity
 import com.google.android.material.snackbar.Snackbar
-import org.sopt.dosopttemplate.ServicePool.authService
+import org.sopt.dosopttemplate.R
+import org.sopt.dosopttemplate.domain.model.RequestSignUpDto
+import org.sopt.dosopttemplate.network.ServicePool.authService
+import org.sopt.dosopttemplate.domain.model.UserInfo
 import org.sopt.dosopttemplate.databinding.ActivitySignupBinding
+import org.sopt.dosopttemplate.util.defaultUserInfo
+import org.sopt.dosopttemplate.util.hideKeyboard
+import org.sopt.dosopttemplate.util.saveAsJsonFile
+import org.sopt.dosopttemplate.util.showSnackbar
+import org.sopt.dosopttemplate.util.toJson
 import retrofit2.Call
 import retrofit2.Callback
 import retrofit2.Response
